fix: backend deployment issues
